Multiple Proportional Density Functions for Continuous Variables

This function gets the proportional density functions for selected distributions against continuous, non-negative numbers. Possible distributions include "normal", "lognormal", "gamma", "exponential", and "all".

Usage

multipdf_cont(var, seq_length = 50, distributions = "all")

Arguments

var

The variable of which to get the PDF.

seq_length

The length of sequence to fit the distribution to

distributions

The distributions to fit x against

Value

A dataframe with x, the real density, and the pdf of the desired distributions with length (nrows) equal to seq_length +1.

Examples

multipdf_cont(iris$Petal.Length)
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#>    var_seq         dens pdf_normal pdf_lognormal  pdf_gamma pdf_exponential
#> 1    1.000 0.0002962162 0.06636660    0.09219444 0.09174455      0.20392849
#> 2    1.118 0.0008659179 0.07355838    0.11830415 0.10877882      0.19762469
#> 3    1.236 0.0023020474 0.08116358    0.14357323 0.12539018      0.19151575
#> 4    1.354 0.0055665834 0.08915315    0.16692167 0.14124545      0.18559564
#> 5    1.472 0.0122442320 0.09748970    0.18763544 0.15608169      0.17985854
#> 6    1.590 0.0245146762 0.10612732    0.20531474 0.16970033      0.17429879
#> 7    1.708 0.0446375243 0.11501174    0.21980734 0.18196024      0.16891089
#> 8    1.826 0.0738667561 0.12408053    0.23114313 0.19277056      0.16368955
#> 9    1.944 0.1110912198 0.13326361    0.23947727 0.20208344      0.15862960
#> 10   2.062 0.1518426552 0.14248397    0.24504429 0.20988722      0.15372607
#> 11   2.180 0.1886234412 0.15165855    0.24812293 0.21620006      0.14897412
#> 12   2.298 0.2129195893 0.16069942    0.24901045 0.22106418      0.14436905
#> 13   2.416 0.2184745880 0.16951501    0.24800454 0.22454070      0.13990634
#> 14   2.534 0.2039872038 0.17801169    0.24539133 0.22670505      0.13558158
#> 15   2.652 0.1737139035 0.18609528    0.24143786 0.22764310      0.13139051
#> 16   2.770 0.1357500773 0.19367283    0.23638790 0.22744774      0.12732899
#> 17   2.888 0.0989395824 0.20065431    0.23046008 0.22621604      0.12339301
#> 18   3.006 0.0701374606 0.20695446    0.22384765 0.22404685      0.11957871
#> 19   3.124 0.0528454118 0.21249443    0.21671936 0.22103887      0.11588231
#> 20   3.242 0.0475961980 0.21720350    0.20922086 0.21728905      0.11230018
#> 21   3.360 0.0531668202 0.22102050    0.20147665 0.21289136      0.10882877
#> 22   3.478 0.0677176686 0.22389519    0.19359200 0.20793580      0.10546468
#> 23   3.596 0.0893953000 0.22578936    0.18565511 0.20250772      0.10220457
#> 24   3.714 0.1163894548 0.22667763    0.17773905 0.19668728      0.09904524
#> 25   3.832 0.1466502199 0.22654805    0.16990366 0.19054916      0.09598357
#> 26   3.950 0.1777611625 0.22540238    0.16219730 0.18416235      0.09301654
#> 27   4.068 0.2069758059 0.22325600    0.15465844 0.17759010      0.09014123
#> 28   4.186 0.2315351920 0.22013761    0.14731709 0.17088997      0.08735480
#> 29   4.304 0.2491254721 0.21608860    0.14019603 0.16411393      0.08465451
#> 30   4.422 0.2582726067 0.21116209    0.13331197 0.15730859      0.08203768
#> 31   4.540 0.2586457397 0.20542179    0.12667647 0.15051535      0.07950174
#> 32   4.658 0.2508848833 0.19894067    0.12029686 0.14377073      0.07704420
#> 33   4.776 0.2362970260 0.19179934    0.11417691 0.13710664      0.07466262
#> 34   4.894 0.2164570173 0.18408446    0.10831747 0.13055065      0.07235466
#> 35   5.012 0.1929309681 0.17588696    0.10271705 0.12412635      0.07011805
#> 36   5.130 0.1672098946 0.16730026    0.09737225 0.11785360      0.06795057
#> 37   5.248 0.1407576343 0.15841857    0.09227813 0.11174890      0.06585010
#> 38   5.366 0.1149788685 0.14933515    0.08742859 0.10582564      0.06381455
#> 39   5.484 0.0910924369 0.14014076    0.08281664 0.10009442      0.06184192
#> 40   5.602 0.0699537299 0.13092223    0.07843458 0.09456330      0.05993028
#> 41   5.720 0.0519814201 0.12176117    0.07427427 0.08923810      0.05807772
#> 42   5.838 0.0372288091 0.11273290    0.07032725 0.08412260      0.05628243
#> 43   5.956 0.0255243329 0.10390562    0.06658487 0.07921884      0.05454264
#> 44   6.074 0.0166097058 0.09533973    0.06303843 0.07452725      0.05285663
#> 45   6.192 0.0101651741 0.08708738    0.05967924 0.07004693      0.05122273
#> 46   6.310 0.0058002361 0.07919232    0.05649869 0.06577579      0.04963934
#> 47   6.428 0.0030645597 0.07168980    0.05348834 0.06171074      0.04810490
#> 48   6.546 0.0014912816 0.06460679    0.05063994 0.05784782      0.04661789
#> 49   6.664 0.0006649257 0.05796228    0.04794546 0.05418237      0.04517685
#> 50   6.782 0.0002706844 0.05176774    0.04539713 0.05070914      0.04378035
#> 51   6.900 0.0001003300 0.04602771    0.04298744 0.04742240      0.04242702

multipdf_cont(iris$Sepal.Length, 100, c("normal", "lognormal"))
#>     var_seq         dens pdf_normal pdf_lognormal
#> 1     4.300 0.0002300378 0.08412707    0.07132138
#> 2     4.336 0.0004204397 0.09119028    0.08000264
#> 3     4.372 0.0007452428 0.09865860    0.08934510
#> 4     4.408 0.0012740603 0.10653567    0.09934885
#> 5     4.444 0.0021149156 0.11482296    0.11000769
#> 6     4.480 0.0033956298 0.12351967    0.12130890
#> 7     4.516 0.0052804924 0.13262248    0.13323298
#> 8     4.552 0.0079753978 0.14212544    0.14575359
#> 9     4.588 0.0116730726 0.15201979    0.15883751
#> 10    4.624 0.0166171237 0.16229387    0.17244474
#> 11    4.660 0.0229956308 0.17293294    0.18652873
#> 12    4.696 0.0309792790 0.18391917    0.20103663
#> 13    4.732 0.0407282850 0.19523151    0.21590971
#> 14    4.768 0.0522777489 0.20684569    0.23108384
#> 15    4.804 0.0656934857 0.21873421    0.24649009
#> 16    4.840 0.0809432606 0.23086632    0.26205534
#> 17    4.876 0.0979851878 0.24320813    0.27770299
#> 18    4.912 0.1168126373 0.25572269    0.29335376
#> 19    4.948 0.1373310675 0.26837007    0.30892643
#> 20    4.984 0.1595019694 0.28110757    0.32433873
#> 21    5.020 0.1831854793 0.29388990    0.33950814
#> 22    5.056 0.2081448051 0.30666938    0.35435273
#> 23    5.092 0.2340189822 0.31939627    0.36879201
#> 24    5.128 0.2602136651 0.33201897    0.38274770
#> 25    5.164 0.2859677731 0.34448444    0.39614453
#> 26    5.200 0.3103549104 0.35673850    0.40891092
#> 27    5.236 0.3324368342 0.36872619    0.42097967
#> 28    5.272 0.3512257112 0.38039224    0.43228853
#> 29    5.308 0.3661467619 0.39168142    0.44278075
#> 30    5.344 0.3768260581 0.40253897    0.45240551
#> 31    5.380 0.3833857602 0.41291108    0.46111830
#> 32    5.416 0.3864300262 0.42274531    0.46888116
#> 33    5.452 0.3867843351 0.43199101    0.47566298
#> 34    5.488 0.3856271638 0.44059977    0.48143951
#> 35    5.524 0.3840354664 0.44852585    0.48619349
#> 36    5.560 0.3829380967 0.45572656    0.48991454
#> 37    5.596 0.3829006395 0.46216265    0.49259911
#> 38    5.632 0.3841429267 0.46779870    0.49425025
#> 39    5.668 0.3864438288 0.47260337    0.49487742
#> 40    5.704 0.3893535531 0.47654978    0.49449613
#> 41    5.740 0.3922770044 0.47961569    0.49312765
#> 42    5.776 0.3946657428 0.48178374    0.49079859
#> 43    5.812 0.3961049635 0.48304161    0.48754047
#> 44    5.848 0.3964610954 0.48338214    0.48338927
#> 45    5.884 0.3958238334 0.48280338    0.47838496
#> 46    5.920 0.3944823234 0.48130864    0.47257102
#> 47    5.956 0.3928408501 0.47890642    0.46599392
#> 48    5.992 0.3912794779 0.47561037    0.45870261
#> 49    6.028 0.3900605629 0.47143912    0.45074807
#> 50    6.064 0.3892540509 0.46641614    0.44218281
#> 51    6.100 0.3887140119 0.46056950    0.43306036
#> 52    6.136 0.3880933514 0.45393161    0.42343488
#> 53    6.172 0.3869059437 0.44653893    0.41336069
#> 54    6.208 0.3846298118 0.43843163    0.40289189
#> 55    6.244 0.3807890531 0.42965323    0.39208196
#> 56    6.280 0.3749745891 0.42025020    0.38098347
#> 57    6.316 0.3670174997 0.41027158    0.36964769
#> 58    6.352 0.3568163493 0.39976851    0.35812440
#> 59    6.388 0.3444790735 0.38879384    0.34646155
#> 60    6.424 0.3302020123 0.37740168    0.33470513
#> 61    6.460 0.3142037346 0.36564693    0.32289894
#> 62    6.496 0.2968204227 0.35358487    0.31108444
#> 63    6.532 0.2783476092 0.34127075    0.29930066
#> 64    6.568 0.2591424082 0.32875936    0.28758410
#> 65    6.604 0.2395770153 0.31610460    0.27596866
#> 66    6.640 0.2200486431 0.30335919    0.26448558
#> 67    6.676 0.2009604461 0.29057427    0.25316349
#> 68    6.712 0.1827131104 0.27779908    0.24202837
#> 69    6.748 0.1656619297 0.26508069    0.23110361
#> 70    6.784 0.1500738495 0.25246376    0.22041002
#> 71    6.820 0.1362021701 0.23999027    0.20996593
#> 72    6.856 0.1241099322 0.22769940    0.19978728
#> 73    6.892 0.1138460512 0.21562731    0.18988764
#> 74    6.928 0.1053449137 0.20380710    0.18027840
#> 75    6.964 0.0984333254 0.19226866    0.17096881
#> 76    7.000 0.0929229351 0.18103867    0.16196614
#> 77    7.036 0.0884719860 0.17014055    0.15327576
#> 78    7.072 0.0847338417 0.15959452    0.14490128
#> 79    7.108 0.0813100240 0.14941760    0.13684470
#> 80    7.144 0.0778128040 0.13962372    0.12910649
#> 81    7.180 0.0738953626 0.13022378    0.12168573
#> 82    7.216 0.0693261240 0.12122579    0.11458027
#> 83    7.252 0.0639876557 0.11263502    0.10778679
#> 84    7.288 0.0579014194 0.10445409    0.10130096
#> 85    7.324 0.0512347642 0.09668323    0.09511752
#> 86    7.360 0.0442291355 0.08932037    0.08923043
#> 87    7.396 0.0371946435 0.08236135    0.08363295
#> 88    7.432 0.0304306539 0.07580016    0.07831772
#> 89    7.468 0.0242066673 0.06962904    0.07327691
#> 90    7.504 0.0186992106 0.06383874    0.06850223
#> 91    7.540 0.0140333962 0.05841870    0.06398508
#> 92    7.576 0.0102204656 0.05335722    0.05971661
#> 93    7.612 0.0072200666 0.04864162    0.05568774
#> 94    7.648 0.0049538262 0.04425849    0.05188929
#> 95    7.684 0.0032918272 0.04019377    0.04831202
#> 96    7.720 0.0021242871 0.03643297    0.04494665
#> 97    7.756 0.0013282586 0.03296128    0.04178395
#> 98    7.792 0.0008038704 0.02976372    0.03881477
#> 99    7.828 0.0004727911 0.02682526    0.03603006
#> 100   7.864 0.0002684891 0.02413095    0.03342093
#> 101   7.900 0.0001480189 0.02166599    0.03097866