multipdf_plot

This function is a version of MultiPDF_plot from my MultiFitR package that is customized to use formatting that I like. However, this formatting uses several packages that I didn't want to include in the base MultiFitR or MultiFitRgg packages.

Usage

multipdf_plot(
  var,
  seq_length = 50,
  distributions = "all",
  palette = "oslo",
  var_name = NULL
)

Arguments

var

The variable to for which to plot PDFs

seq_length

The number of points over which to fit x

distributions

The distributions to fit x against

palette

The color palette to use on the graph

var_name

The variable name to use for x. If no name is provided, the function will grab the column name provided in x

Value

A plot showing the PDF of the selected variable against the selected distributions over the selected sequence length